requirements_base.txt
otree>=2.0.27
django-redis==4.9.0
websocket-client==0.47.0
numpy==1.13.3
websocket-client==0.47.0